move data source into src
