+++ backend/package.json
@@ -6,7 +6,7 @@
   "private": true,
   "license": "MIT",
   "scripts": {
-    "build": "nest build",
+    "build": "npx tsc -p ../shared/tsconfig.json && npx tsc -p tsconfig.build.json",
     "format": "prettier --write \"src/**/*.ts\" \"test/**/*.ts\"",
     "start": "node dist/main.js",
     "start:dev": "cross-env NODE_ENV=development nest start --watch",
@@ -21,8 +21,8 @@
     "test:watch": "cross-env NODE_ENV=test jest --watch",
     "migration:create": "ts-node --transpile-only ./node_modules/typeorm/cli.js migration:create",
     "migration:generate": "node scripts/generate-migration.mjs",
-    "migration:run": "cross-env NODE_ENV=development ts-node --transpile-only ./node_modules/typeorm/cli.js migration:run -d data-source.ts",
-    "migration:revert": "ts-node --transpile-only ./node_modules/typeorm/cli.js migration:revert -d data-source.ts",
+    "migration:run": "cross-env NODE_ENV=development ts-node --transpile-only ./node_modules/typeorm/cli.js migration:run -d src/data-source.ts",
+    "migration:revert": "ts-node --transpile-only ./node_modules/typeorm/cli.js migration:revert -d src/data-source.ts",
     "migration:create:dev": "cross-env NODE_ENV=development npm run migration:create --",
     "migration:generate:dev": "cross-env NODE_ENV=development npm run migration:generate --",
     "migration:revert:dev": "cross-env NODE_ENV=development npm run migration:revert --",
