+++ backend/src/data-source.ts
@@ -2,7 +2,7 @@ import { config } from 'dotenv';
 config({ path: `.env.${process.env.NODE_ENV || 'development'}` });
 import { DataSource } from 'typeorm';
 import { join } from 'path';
-import { getCurrentCompanyId } from './src/common/tenant/tenant-context';
+import { getCurrentCompanyId } from './common/tenant/tenant-context';
 
 const isProduction = process.env.NODE_ENV === 'production';
 
@@ -14,7 +14,7 @@ const dataSource = new DataSource({
   password: process.env.DB_PASSWORD,
   database: process.env.DB_NAME,
   entities: [join(__dirname, '**/*.entity{.ts,.js}')],
-  migrations: [join(__dirname, 'src/migrations/*{.ts,.js}')],
+  migrations: [join(__dirname, 'migrations/*{.ts,.js}')],
   migrationsRun: process.env.RUN_MIGRATIONS === 'true',
   synchronize: false,
   ssl: isProduction ? { rejectUnauthorized: false } : false,
