+++ backend/tsconfig.build.json
@@ -2,12 +2,17 @@
   "extends": "./tsconfig.json",
   "compilerOptions": {
     "outDir": "./dist",
+    "rootDir": "src",
     "declaration": true,
     "sourceMap": true,
     "removeComments": true,
     "emitDecoratorMetadata": true,
     "experimentalDecorators": true,
-    "incremental": true
+    "incremental": true,
+    "paths": {
+      "@rflp/shared": ["../shared/dist"],
+      "@rflp/shared/*": ["../shared/dist/*"]
+    }
   },
   "exclude": ["node_modules", "test", "dist", "**/*.spec.ts"]
 }
