+++ docker-compose.yml
@@ -46,6 +46,7 @@ services:
       - ./backend/.env.development
     environment:
       - NODE_ENV=development
+      - CHOKIDAR_USEPOLLING=true
     ports:
       - "3000:3000"
     networks:
@@ -65,6 +66,7 @@ services:
       - "4200:4200"
     environment:
       - NODE_ENV=development
+      - CHOKIDAR_USEPOLLING=true
     networks:
       - rflandscaperpro
